+++ nautobot_ssot_aristacv/tests/test_utils_cloudvision.py
@@ -108,7 +108,7 @@ class TestCloudvisionUtils(TestCase):
         mock_tag.value.creator_type = 1
 
         device_tag_stub = MagicMock()
-        device_tag_stub.DeviceTagServiceStub.return_value.GetAll.return_value = [mock_tag]
+        device_tag_stub.TagServiceStub.return_value.GetAll.return_value = [mock_tag]
 
         with patch("nautobot_ssot_aristacv.utils.cloudvision.tag_services", device_tag_stub):
             results = cloudvision.get_tags_by_type(client=self.client)
@@ -123,7 +123,7 @@ class TestCloudvisionUtils(TestCase):
         mock_tag.value.device_id.value = "JPE12345678"
 
         tag_stub = MagicMock()
-        tag_stub.DeviceTagAssignmentConfigServiceStub.return_value.GetAll.return_value = [mock_tag]
+        tag_stub.TagAssignmentConfigServiceStub.return_value.GetAll.return_value = [mock_tag]
 
         with patch("nautobot_ssot_aristacv.utils.cloudvision.tag_services", tag_stub):
             results = cloudvision.get_device_tags(client=self.client, device_id="JPE12345678")
@@ -194,7 +194,6 @@ class TestCloudvisionUtils(TestCase):
 
     def test_get_interfaces_chassis(self):
         """Test get_interfaces_chassis method."""
-        self.maxDiff = None  # pylint:disable=invalid-name
         mock_query = MagicMock()
         mock_query.dataset.type = "device"
         mock_query.dataset.name = "JPE12345678"
@@ -263,6 +262,46 @@ class TestCloudvisionUtils(TestCase):
                 client=self.client, dId="JPE12345678", interface="Ethernet1"
             )
         self.assertEqual(results, "xcvr1000BaseT")
+
+    def test_get_interface_mode_trunk(self):
+        """Test the get_interface_mode method for a trunk."""
+        mock_query = MagicMock()
+        mock_query.dataset.type = "device"
+        mock_query.dataset.name = "JPE12345678"
+        mock_query.paths.path_elements = [
+            "\304\005Sysdb",
+            "\304\010bridging",
+            "\304\020switchIntfConfig",
+            "\304\020switchIntfConfig",
+            "\304\tEthernet1",
+        ]
+
+        with patch("cloudvision.Connector.grpc_client.grpcClient.create_query", mock_query):
+            self.client.get = MagicMock()
+            self.client.get.return_value = fixtures.TRUNK_INTF_MODE_QUERY
+            results = cloudvision.get_interface_mode(client=self.client, dId="JPE12345678", interface="Ethernet1")
+        expected = "trunk"
+        self.assertEqual(results, expected)
+
+    def test_get_interface_mode_access(self):
+        """Test the get_interface_mode method for a access."""
+        mock_query = MagicMock()
+        mock_query.dataset.type = "device"
+        mock_query.dataset.name = "JPE12345678"
+        mock_query.paths.path_elements = [
+            "\304\005Sysdb",
+            "\304\010bridging",
+            "\304\020switchIntfConfig",
+            "\304\020switchIntfConfig",
+            "\304\tEthernet5",
+        ]
+
+        with patch("cloudvision.Connector.grpc_client.grpcClient.create_query", mock_query):
+            self.client.get = MagicMock()
+            self.client.get.return_value = fixtures.ACCESS_INTF_MODE_QUERY
+            results = cloudvision.get_interface_mode(client=self.client, dId="JPE12345678", interface="Ethernet5")
+        expected = "access"
+        self.assertEqual(results, expected)
 
     port_types = [
         ("built_in_gig", {"port_info": {}, "transceiver": "xcvr1000BaseT"}, "1000base-t"),